refactor: change file names and add dailyCard file
